requirements.txt
Pillow>=9.0.0
easyocr>=1.6.0
numpy>=1.21.0
# tkinter is included with Python by default
